Split queries components out
